+++ data-raw/DATASET.R
@@ -4,11 +4,15 @@
 
 md <- pipload::pip_load_data("PRY", 2015)
 md_ex <- md[1:10]
-usethis::use_data(md_ex, overwrite = TRUE, internal = TRUE)
+
+readr::write_rds(x    = md_ex,
+                 file = "tests/testdata/md_ex.rds")
 
 gd <- pipload::pip_load_data("CHN", 2015)
 gd_ex <- gd[1:10]
-usethis::use_data(gd_ex, overwrite = TRUE, internal = TRUE)
+
+readr::write_rds(x    = gd_ex,
+                 file = "tests/testdata/gd_ex.rds")