+++ NEWS.md
@@ -6,7 +6,7 @@
 
 # pipload 0.1.15
 
-* Change creation of pip clases in S3 mode. now, the main function is `as_pip`
+* Change creation of pip classes in S3 mode. now, the main function is `as_pip`
 
 # pipload 0.1.14